libcalendar weekday(): a subtraction that overflows for the bottom 729,652 of its int domain
weekday() caches the day number of a known Monday -- 729652 -- and does
`nd = (nd - nmonday) % 7'.  For every nd below INT_MIN + 729652 that
subtraction overflows, which is undefined rather than merely wrong.  It
is a public libcalendar entry point with no stated domain, so that is
its whole int range; the only in-tree caller passes an ndaysgi() result
and stays well inside it, which is why nothing had noticed.  Widened to
long long, which holds every int difference exactly and still fits the
result in an int, so nothing in range changes.

Found by the fuzzing engine rather than the analyser or the model
checker: AFL produced nd = 0x8000550b and the replay came back
"undefined-behavior calendar.c:283:11".  CLEAN on the same budget after.

THREE OF THE FIVE CRASHES THAT RUN REPORTED WERE THE INSTRUMENT

4,390 pairs over lib/libc and lib/msun.  abs and imaxabs are real and
stand (the documented INT_MIN UB), weekday is real and is fixed above.
The other two were not the code:

  valloc(0x8700000000) -- ASan's allocator ABORTS on a request it
  considers absurd, and absurd is exactly what a fuzzer feeds a function
  whose only parameter is a size.  The C library returns NULL with
  ENOMEM.  allocator_may_return_null=1 restores that, and a function
  that then USES the NULL still crashes, which is the finding worth
  having.  CLEAN with it set.

  memalign(1 << 63, 0) -- "AddressSanitizer: CHECK failed:
  asan_allocator.cpp:601" is ASan's own arithmetic overflowing and says
  nothing about the code beneath it.  New SANFAIL status rather than a
  drop: it is a real limit on what this engine can see, and a limit
  nobody counts is a limit nobody knows about.

Classifying it needed one more fix.  _tail() is right for a compiler,
which puts its verdict last, and exactly wrong for a sanitizer, which
puts the verdict FIRST and forty frames after it -- so the first attempt
kept 600 characters of frame numbers and memalign came back CRASH again
with the classifying line three screens above the cut.  A fuzzer-found
CRASH also recorded only AFL's input and an empty detail; it is now
replayed under the sanitizers and the report kept.

THE INCLUDE PATH THE TWO HALVES DO NOT SHARE

cbmc was being run with no -I at all.  854 of the 4,390 came back NOSEED
and the top reasons were "namespace.h: No such file or directory" (201),
"sys/_types.h" (137) and "fpmath.h" (112) -- all three in this tree.
classify.py has computed the right answer per file since the beginning,
so it is asked for the same way rather than approximated a third time.
cbmc is not a compiler driver (it takes -I and -D and answers anything
else with its usage text, which the engine was recording as the reason
for NOSEED), and `-include foo.h' is two argv elements, so the filter
keeps its pairs whole.  NOSEED 854 -> 240.

What it did NOT do: ERROR rose 3412 -> 4026 by nearly the same amount
and CLEAN went 19 -> 21.  The 614 pairs cbmc can now read fail one stage
later, at the harness compile, on the same headers -- a better
diagnosis, not more fuzzing, and the write-up says which.  The harness
keeps the HOST include path deliberately: cbmc reads the code, the
fuzzer runs it on Linux against glibc.  Compiling the unit with the
build's flags was tried and those carry
--target=x86_64-unknown-freebsd15.0; valloc(0) and weekday(729652) then
SEGV'd before executing a line of their own code.  Adding the tree's -I
without the target is no better -- those paths shadow the host's, and
abs and weekday stop linking.

A widening that loses a finding is not a widening: the include path also
hands cbmc's C front end headers it cannot parse, and imaxabs.cpp went
from a real counterexample to "parse error before '__char16_t'".  So the
libc++ shim is dropped for a landed .cpp read as C, and where the
flagged run cannot parse at all the bare run is tried and whichever
answers is kept.  imaxabs is back to CRASH on the real
imaxabs(INTMAX_MIN).

  before   CLEAN 19  CRASH 5  SANFAIL 0  NOSEED 854  ERROR 3412
  after    CLEAN 21  CRASH 2  SANFAIL 1  NOSEED 240  ERROR 4026

with NOFUNC 89 and NORETURN 11 on both sides.

+++ hbsd/src/lib/libcalendar/calendar.c
@@ -279,8 +279,16 @@ weekday(int nd)
 	if (nmonday == 0)
 		nmonday = ndaysgi(&dmondaygi);
 
-	/* return (nd - nmonday) modulo 7 which is the weekday */
-	nd = (nd - nmonday) % 7;
+	/*
+	 * return (nd - nmonday) modulo 7 which is the weekday.
+	 *
+	 * In long long, because nd is this function's whole int domain and
+	 * nmonday is 729652: the subtraction overflows for every nd below
+	 * INT_MIN + 729652, which is undefined rather than merely wrong.
+	 * The wider type is exact for every int and the result still fits
+	 * one, so nothing in range changes.
+	 */
+	nd = (int)(((long long)nd - nmonday) % 7);
 	if (nd < 0)
 		return (nd + 7);
 	else
